+++ src/components/Head.js
@@ -1,0 +1,38 @@
+import * as React from 'react';
+
+export const Head = () => (
+	<>
+		<html lang="en" />
+		<title>ml5.js | Friendly Machine Learning for the Web</title>
+		<meta charset="UTF-8" />
+		<meta
+			name="description"
+			content="ml5.js aims to make machine learning approachable for a broad audience of artists, creative coders, and students. The library provides access to machine learning algorithms and models in the browser, building on top of TensorFlow.js with no other external dependencies."
+		/>
+		<meta
+			property="og:image"
+			content="https://ml5js.org/favicon/og-image.png"
+		/>
+		<meta property="og:image:width" content="1280" />
+		<meta property="og:image:width" content="640" />
+		<link rel="icon" type="image/x-icon" href="/favicon/favicon.ico" />
+		<link
+			rel="apple-touch-icon"
+			sizes="180x180"
+			href="/favicon/apple-touch-icon.png"
+		/>
+		<link
+			rel="icon"
+			type="image/png"
+			sizes="32x32"
+			href="/favicon/favicon-32x32.png"
+		/>
+		<link
+			rel="icon"
+			type="image/png"
+			sizes="16x16"
+			href="/favicon/favicon-16x16.png"
+		/>
+		<link rel="manifest" href="/favicon/site.webmanifest" />
+	</>
+);
