+++ src/layout/Layout.js
@@ -26,4 +26,6 @@ const Layout = ({ children }) => {
 	);
 }
 
+
+
 export default Layout
