+++ src/pages/community.js
@@ -93,7 +93,7 @@ const CommunityPage = ({ data }) => {
   )
 }
 
-export const Head = () => <title>ml5.js | Friendly Machine Learning for the Web</title>
+export { Head } from '../components/Head';
 
 export default CommunityPage;
