+++ src/pages/learn.js
@@ -16,6 +16,5 @@ const LearnPage = () => {
   )
 }
 
-export const Head = () => <title>ml5.js | Friendly Machine Learning for the Web</title>
-
+export { Head } from '../components/Head';
 export default LearnPage
